requirements.txt
opencv-contrib-python==4.8.1.78
numpy<2.0.0
argparse

# Optional dependencies for AI text detection
# Install these for state-of-the-art object detection:
# pip install torch torchvision transformers groundingdino-py supervision

torch>=1.9.0
torchvision>=0.10.0  
transformers>=4.20.0
groundingdino-py>=0.1.0
supervision==0.6.0
pillow>=8.0.0
